requirements.txt
requests==2.31.0       # Para fazer chamadas à API (HTTP Client)
pandas==2.2.0          # Para manipulação tabular (DataFrame)
python-dotenv==1.0.0   # Para ler variáveis do arquivo .env
boto3==1.34.0          # SDK da AWS (usaremos na Etapa 4, mas já instalamos agora)
pyarrow==15.0.0        # Engine para criar arquivos Parquet (mais eficiente que CSV)